Change file size units to IEC 80000-13:2025 (binary multiples) (#2062)
* Change file size units from SI to IEC 80000-13:2025

Increase left-padded (%S) file size string length to 5 characters


* Add note about used IEC units for file size in docs

---------

+++ df_statfs.go
@@ -17,5 +17,5 @@ func diskFree(wd string) string {
 	}
 
 	// Available blocks * size per block = available space in bytes
-	return "df: " + humanize(int64(uint64(stat.Bavail)*uint64(stat.Bsize)))
+	return "df: " + humanize(uint64(stat.Bavail)*uint64(stat.Bsize))
 }
